Created MRI and JRuby specific implementations of CachedThreadPool
* Created new RubyCachedThreadPool class and associated files.
* CachedThreadPool and JavaCachedThreadPool both map to RubyCachedThreadPool.
* Fixed thread pool spec that worked for fixed pools but not cached pools.
* Updated RubyCachedThreadPool constructor to be more like RubyFixedThreadPool.
* Finished JRuby optimized CachedThreadPool.
